Instrument & vocal track refactors around special phrases
Instead of collecting all specialphrases under a single list, this refactor splits the list between each of the phrases that a track can make use of.

Instrument/Difficulty tracks: Overdrives, Soloes, Trills, Tremolos, BREs, and Faceoffs(1 & 2).
VocalTrack: Overdrives, VocalPhrases(1 & 2), HarmonyLines, RangeShifts, and LyricShifts.

ProGuitarInstrumentTrack sees the same patten for the ChordPhrases list.

Loading charts from a midi file benefits substantially from this change. Instead of having to pre-add a node at position and traversing backwards to add a phrase when a noteOff occurs, we can now just instantly append a phrase to the end of it's specific list.

Other updates:
+ Implement a PhraseTrack base type for Instrument/Difficultytrack. This makes the interface for DrumTrack conversion simpler to manage (but also opens the door for future use cases).
+ Removed VocalPercussion2 type. It was literally just a boolean anyways.
+ Don't *always* trim excess note buffer data.